Desktop/IIT B Assignment/total-IITB-project/courses_api/requirements.txt
Django==3.2.5
djangorestframework==3.12.4
psycopg2-binary==2.9.3